requirements.txt
fastapi
uvicorn
requests
slowapi
google-generativeai
python-dotenv